requirements.txt
requests
beautifulsoup4
lxml
langchain
langchain-openai
langchain-community
langchain-text-splitters
openai
faiss-cpu
streamlit
sentence-transformers
python-dotenv
langchain-google-genai